wardriving_analyzer_v4.4.3_playback_docked/wardriving_analyzer/backend/requirements.txt
# Wardriving Analyzer Requirements
# Compatible with Python 3.10, 3.11, 3.12

# Web Framework
Flask>=3.0.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0
Flask-SocketIO>=5.3.0,<6.0.0
python-socketio>=5.9.0,<6.0.0
python-engineio>=4.7.0,<5.0.0
Werkzeug>=3.0.0,<4.0.0

# Data Analysis (versions with pre-built wheels)
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# MQTT (optional - for live data)
paho-mqtt>=1.6.0,<2.0.0

# PDF Generation
reportlab>=4.0.0,<5.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
